+++ pypeit/par/util.py
@@ -476,7 +476,7 @@ def _parse_setup_lines(lines):
     return [ l.split()[1].strip() for l in lines if 'Setup' in l ]
 
 
-def parse_pypeit_file(ifile, file_check=True):
+def parse_pypeit_file(ifile, file_check=True, runtime=False):
     """
     Parse the user-provided .pypeit reduction file.
 
@@ -486,6 +486,8 @@ def parse_pypeit_file(ifile, file_check=True):
         file_check (:obj:`bool`, optional):
             Check that the files in the pypeit configuration data file
             exist, and fault if they do not.
+        runtime (:obj:`boo`, optional):
+            Perform additional checks if called to run PypeIt
 
     Returns:
         :obj:`lists`: Four lists are provided:
@@ -519,6 +521,7 @@ def parse_pypeit_file(ifile, file_check=True):
         data_files, frametype, usrtbl = _read_data_file_table(lines[s:e], file_check=file_check)
     is_config[s-1:e+1] = False
     if len(data_files) == 0 and file_check:
+        import pdb; pdb.set_trace()
         msgs.error('There are no raw data frames' + msgs.newline() +
                    'Perhaps the path to the data is incorrect?')
     else:
@@ -533,6 +536,15 @@ def parse_pypeit_file(ifile, file_check=True):
     else:
         setups = _parse_setup_lines(lines[s:e])
         is_config[s-1:e+1] = False
+
+    # Running PypeIt?
+    if runtime:
+        for key in ['filename', 'frametype']:
+            if key not in usrtbl.keys():
+                msgs.error("Need to add {:s} to your PypeIt file before using run_pypeit".format(key))
+        # Setup
+        if len(setups) != 1:
+            msgs.error("Need to add setup info to your PypeIt file in the setup block!")
 
     msgs.info('Input file loaded successfully')
     return list(lines[is_config]), data_files, frametype, usrtbl, setups